+++ cw-vault-standard/src/extensions/lockup.rs
@@ -24,6 +24,11 @@ pub enum LockupExecuteMsg {
     /// CW4626 and native tokens.
     Unlock {
         /// The amount of vault tokens to unlock.
+        #[deprecated(
+            since = "0.4.1",
+            note = "This field will be removed in the next version. The amount \
+            of vault tokens should instead be read from the actual amount of sent vault tokens."
+        )]
         amount: Uint128,
     },