+++ cw-vault-standard/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deprecated `VaultStandardQueryMsg::PreviewDeposit` and `VaultStandardQueryMsg::PreviewRedeem`.
     - There queries turned out to be too difficult to implement in most cases. We recommend to use transaction simulation from non-contract clients such as frontends. These queries will be removed in the next major version.
 - Deprecated CW4626 Extension now that TokenFactory module is standard across all major Cosmos chains. The extension will be removed in the next major version.
+- Deprecated `amount` fields in `VaultStandardExecuteMsg::Deposit` and `VaultStandardExecuteMsg::Redeem` as well as in `LockupExecuteMsg::Unlock` and `ForceUnlockExecuteMsg::ForceRedeem`. These fields will be removed in a future version and the amount from the sent funds will be used instead.
 
 ## [0.3.3] - 2023-09-27